% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATC_parameters.R
\docType{data}
\name{ATC_parameters}
\alias{ATC_parameters}
\title{Parameters for ATC Codes}
\format{
A `data.frame` or `data.table` with one row per ATC code. Includes the following variables:
\describe{
  \item{partial_atc}{Partial or full ATC code (character)}
  \item{lower_ddd_atc}{Minimum commonly used daily dose (numeric, in DDDs)}
  \item{usual_ddd_atc}{Usual daily dose (numeric, in DDDs)}
  \item{minimum_dur_atc}{Minimum allowed treatment duration (numeric, in days)}
  \item{maximum_dur_atc}{Maximum allowed treatment duration (numeric, in days)}
}
}
\source{
Parameter set used by PRE2DUP-R.
}
\usage{
ATC_parameters
}
\description{
A dataset containing parameters associated with ATC (Anatomical Therapeutic Chemical) codes used in the PRE2DUP-R package.
}
\details{
The ATC codes may be partial or complete; the program uses the closest match to the ATC code in the drug purchases. All ATC codes appearing in drug purchases must be included in the ATC parameters table, at least at a partial-code level. These parameters are used when package-specific information is not available.


See more about ATC classification at \url{https://atcddd.fhi.no/atc/structure_and_principles/} and about DDDs at \url{https://atcddd.fhi.no/ddd/definition_and_general_considera/}.
}
\seealso{
\code{\link{pre2dup}}, \code{\link{check_atc_parameters}}
}
\keyword{datasets}
