% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dup_hosp_days.R
\name{calc_dup_hosp_days}
\alias{calc_dup_hosp_days}
\title{Calculate Hospital Days in Drug Use Periods}
\usage{
calc_dup_hosp_days(
  period_no,
  pre_id,
  dup_start,
  dup_end,
  hosp_id,
  hosp_in,
  hosp_out
)
}
\arguments{
\item{period_no}{Integer vector for period identification.}

\item{pre_id}{Vector of person identifiers for drug purchases.}

\item{dup_start}{Integer vector for the start of drug use periods.}

\item{dup_end}{Integer vector for the end of drug use periods.}

\item{hosp_id}{Vector of person identifiers for hospitalizations.}

\item{hosp_in}{Integer vector for hospitalization admission dates.}

\item{hosp_out}{Integer vector for hospitalization discharge dates.}
}
\value{
A data.table with two columns:
\describe{
  \item{period}{Period identifier.}
  \item{dup_hospital_days}{Number of hospitalization days during the period. Discharge dates are not included in the hospital stay.}
}
}
\description{
Calculates the total number of hospital days during drug use periods. This function is for internal use of PRE2DUP-R.
}
\keyword{internal}
