% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_arguments}
\alias{check_arguments}
\title{Checks Missing or Unvalid Arguments in Function Calls}
\usage{
check_arguments(dt, required_columns, data_name = NULL, ...)
}
\arguments{
\item{dt}{required argument in the function call of any the datasets.}

\item{required_columns}{a character vector of required column names that must be present in the function call.}

\item{data_name}{required argument in the function call of any the datasets.}

\item{...}{other arguments passed to the function call.}
}
\description{
This function checks if the required arguments are present in the function call. It compares the given arguments to main function argument list. It raises an error if any required arguments are missing, NULL, or empty strings and stops execution of check function. It is used used internally in the package to ensure that the necessary arguments are provided when calling the check functions.
}
\keyword{internal}
