% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_coverage}
\alias{check_coverage}
\title{Check Coverage of Values by ATC Codes}
\usage{
check_coverage(ATC, val, limit = 90, opti = c("missing", "zero"))
}
\arguments{
\item{ATC}{a character vector of ATC codes.}

\item{val}{a numeric vector of values corresponding to the ATC codes.}

\item{limit}{a numeric value representing the upper limit for coverage percentage. Default is 90.}

\item{opti}{a character string indicating the type of coverage check to perform. It can be either "missing" (to check for non-missing values) or "zero" (to check for values greater than zero). Default is "missing".}
}
\value{
a character string listing ATC codes that do not meet the coverage limit, formatted as "ATC (coverage%)". If all ATC codes meet the limit, returns NULL.
}
\description{
Check Coverage of Values by ATC Codes
}
\keyword{internal}
