% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_date_range}
\alias{check_date_range}
\title{Check Date Range}
\usage{
check_date_range(date_range)
}
\arguments{
\item{date_range}{a character vector of length 2 containing the start and end dates.}
}
\description{
Checks if the provided date range is valid. It ensures that the date range contains exactly two valid dates and that the starting date is before the ending date. It is used internally in the package to validate date ranges provided by the user.
}
\keyword{internal}
