% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_numeric}
\alias{check_numeric}
\title{Check Numeric Values}
\usage{
check_numeric(col, allow_na = FALSE, allow_zero = FALSE)
}
\arguments{
\item{col}{a numeric vector to be checked for invalid values.}

\item{allow_na}{logical value indicating whether NA values are allowed. Default is FALSE.}

\item{allow_zero}{logical value indicating whether zero values are allowed. Default is FALSE.}
}
\value{
a vector of indices where the numeric values are invalid (i.e., infinite, negative, NA, or zero if not allowed).
}
\description{
This function checks a numeric vector for invalid values. It identifies indices of elements that are infinite, negative, NA (if not allowed), or zero (if not allowed). It is used internally in the package to validate numeric columns that are expected to contain valid non-negative numbers.
}
\keyword{internal}
