% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_order}
\alias{check_order}
\title{Check Order of Values}
\usage{
check_order(lower, usual = NULL, upper)
}
\arguments{
\item{lower}{a numeric vector representing the lower limit.}

\item{usual}{a numeric vector representing the usual value (optional).}

\item{upper}{a numeric vector representing the upper limit.}
}
\value{
a vector of indices where the order of values is incorrect (i.e., lower > upper, lower > usual, or usual > upper).
}
\description{
This function checks the order of maximum of three numeric vectors: lower, usual, and upper. It returns the indices where the order is incorrect (i.e., lower > upper, lower > usual, or usual > upper). It is used internally in the package to validate the order of values in ATC and package parameters.
}
\keyword{internal}
