% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{combine_overlaps}
\alias{combine_overlaps}
\title{Combine Overlapping Hospitalizations}
\usage{
combine_overlaps(personid, admission, discharge)
}
\arguments{
\item{personid}{a vector of person identifiers in hospitalization data.}

\item{admission}{a vector of admission dates.}

\item{discharge}{a vector of discharge dates.}
}
\value{
a data.table with combined hospitalizations, where overlapping hospitalizations are merged into single records.
}
\description{
Combine Overlapping Hospitalizations
}
\keyword{internal}
