% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{count_sliding_average}
\alias{count_sliding_average}
\title{Count sliding average}
\usage{
count_sliding_average(
  ddd,
  time,
  w_past = 1,
  w_present = 4,
  w_future = 1,
  w_limit = 5,
  max_refill = 300
)
}
\arguments{
\item{ddd}{Amount of DDDs purchased}

\item{time}{Time from previous purchase. For first purchase 0.}

\item{w_past}{Weight for purchase before current purchase.}

\item{w_present}{Weight for current purchase.}

\item{w_future}{Weight for purchase after current purchase.}

\item{w_limit}{Weight for first and last purchase.}

\item{max_refill}{Maximum time allowed between purchases}
}
\value{
This function returns a number vector
with sliding averages
}
\description{
This function counts sliding average of Defined Daily
Dose (DDD) from puchased DDDs and times between purchases.
}
\examples{
\dontrun{
ddds <- c(15, 15, 15, 15, 15)
times <- c(0, 59, 54, 69, 51)
count_sliding_average(ddds, times)
}


}
