% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{err_message}
\alias{err_message}
\title{Creation of Error Message}
\usage{
err_message(miss, error, arg_or_col = c("argument", "column"))
}
\arguments{
\item{miss}{a character vector of missing arguments or columns.}

\item{error}{a character string describing the error.}

\item{arg_or_col}{a character string indicating whether the missing items are arguments or columns. It can be either "argument" or "column". The default is "argument"}
}
\value{
a character string containing the formatted error message.
}
\description{
This function creates an error message for missing arguments or columns in the function call. It is used internally in the package to provide informative error messages when required arguments or columns are not provided.
}
\keyword{internal}
