% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{extract_atc_codes}
\alias{extract_atc_codes}
\title{Extract ATC Codes from Text}
\usage{
extract_atc_codes(text)
}
\arguments{
\item{text}{a character vector containing text from which ATC codes need to be extracted. The expected format is "ATC code (percentage)".}
}
\value{
a character vector of extracted ATC codes. If no matches are found, it returns an empty character vector.
}
\description{
This function extracts ATC codes from a character vector containing text formatted as "ATC code (percentage)". It uses a regular expression to find the ATC codes before the parentheses. Function is for internal use for creating error messages in the package.
}
\keyword{internal}
