% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{find_multiple_atcs}
\alias{find_multiple_atcs}
\title{Find Multiple ATC Codes}
\usage{
find_multiple_atcs(atc_col, package_col, location)
}
\arguments{
\item{atc_col}{a character vector containing ATC codes.}

\item{package_col}{an integer vector containing package identifiers corresponding to the ATC codes.}

\item{location}{character. Short description of finding location.}
}
\value{
invisible NULL if all packages have a single ATC code, otherwise raises an error with a message listing the packages that have multiple ATC codes.
}
\description{
This function checks if each package has a single ATC code. It counts the unique ATC codes for each package and raises an error if any package has more than one ATC code. It is used internally in the package to ensure data integrity. Function is for internal use for creating error messages in the package.
}
\keyword{internal}
