% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{make_block}
\alias{make_block}
\title{Make block}
\usage{
make_block(btw, max_length = 300)
}
\arguments{
\item{btw}{Time between purchases.}

\item{max_length}{Maximum allowed time between purchases.}
}
\value{
This function returns a vector of numbers
}
\description{
This function makes blocks of drug purchases based on
maximum allowed time between purchases
}
\note{
For the first purchase time between purchases should be 0.
}
\examples{
\dontrun{
times <- c(0, 31, 44, 22, 110, 350, 33, 42, 54, 100, 500, 24, 33, 37)
make_block(times)
}


}
