% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{make_warning}
\alias{make_warning}
\title{Make Warning Message}
\usage{
make_warning(errorws, colvars, warning_message, print_all = TRUE)
}
\arguments{
\item{errorws}{a vector of row numbers where the warning applies.}

\item{colvars}{a character vector of column names that are involved in the warning.}

\item{warning_message}{a character string containing the warning message to be included in the output.}

\item{print_all}{a logical value indicating whether to print all rows with warnings or truncate the output to the first 5 rows. Default is TRUE.}
}
\description{
This function creates a warning message for a specific condition in the data. It formats the message to include the column names involved, the number of rows affected, and the specific row numbers where the warning applies. It is used internally in the package to provide informative warnings when certain conditions are met.
}
\keyword{internal}
