% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_calculation.R
\name{mode_calculation}
\alias{mode_calculation}
\title{Calculates the most common duration between purchases}
\usage{
mode_calculation(pack_id, pre_ratio, dur, report_iters = 50)
}
\arguments{
\item{pack_id}{vector of package identifiers}

\item{pre_ratio}{vector of package ratios, i.e. number of packages because partial packages are not allowed}

\item{dur}{vector of times between purchases}
}
\value{
data.table with the most common duration between purchases
}
\description{
Calculates the mode of time intervals between purchases for each package.
Intervals within 2 days are grouped together.
The most common (modal) interval is selected from the grouped values.
If multiple modes exist, the one closest to the median is chosen;
if there's still a tie, the one closest to the mean is selected.
Intended for internal use in the PRE2DUP-R package.
}
\keyword{internal}
