% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{PRE2DUPR-package}
\alias{PRE2DUPR}
\alias{PRE2DUPR-package}
\title{Package-level imports}
\description{
PRE2DUP-R for creating drug treatment periods from drug purchases.
}
\author{
\strong{Maintainer}: Pia Vattulainen \email{vattpia@gmail.com}

}
\keyword{internal}
